+++ include/GafferSceneUI/TransformTool.h
@@ -172,6 +172,7 @@ class GAFFERSCENEUI_API TransformTool : public GafferSceneUI::SelectionTool
 				void initFromSceneNode( const GafferScene::SceneAlgo::History *history );
 				void initWalk( const GafferScene::SceneAlgo::History *history, bool &editScopeFound );
 				void throwIfNotEditable() const;
+				Imath::M44f transformToLocalSpace() const;
 
 				GafferScene::ConstScenePlugPtr m_scene;
 				GafferScene::ScenePlug::ScenePath m_path;
@@ -186,6 +187,7 @@ class GAFFERSCENEUI_API TransformTool : public GafferSceneUI::SelectionTool
 				Gaffer::EditScopePtr m_editScope;
 				mutable boost::optional<TransformEdit> m_transformEdit;
 				Imath::M44f m_transformSpace;
+				bool m_aimConstraint;
 
 		};